Extract update hall for to a separte component
